target point manager renamed to target point service
