+++ app/src/main/java/com/rootstrap/android/network/services/ITargetPointService.kt
@@ -1,11 +1,11 @@
-package com.rootstrap.android.network.managers
+package com.rootstrap.android.network.services
 
 import com.rootstrap.android.network.models.Target
 import com.rootstrap.android.network.models.TargetPointSerializer
 import com.rootstrap.android.network.models.TopicsSerializer
 import com.rootstrap.android.util.extensions.Data
 
-interface ITargetPointManager {
+interface ITargetPointService {
 
     suspend fun createTarget(target: Target): Result<Data<TargetPointSerializer>>
