+++ app/src/main/res/layout/activity_target_points.xml
